More warning work.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1103 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/warnings2/kernel/runtime/RexxCore.h
@@ -786,7 +786,7 @@ const char *mempbrk(const char *, const char *, size_t);     /* search for chara
 void report_halt(RexxString *description);
 void report_nomethod(RexxString *message, RexxObject *receiver);
 void missing_argument(LONG position);
-long message_number(RexxString *);
+int  message_number(RexxString *);
 RexxActivity *activity_find (void);
 
 #define report_exception(error) (CurrentActivity->reportAnException(error))
